[BAU] refactor core
