+++ core/exceptions.py
@@ -1,0 +1,5 @@
+class ValidationError(RuntimeError):
+    """Validation error."""
+
+    def __init__(self, validation_failures):
+        self.validation_failures = validation_failures
